Adjust file interface for random access files.

+++ drivers/bootcons.c
@@ -2,7 +2,7 @@
 
 extern void BootConsPutChar(char c);
 
-static int BootConsWrite(__unused File_t *f, const char *buf, size_t nbyte) {
+static long BootConsWrite(__unused File_t *f, const char *buf, size_t nbyte) {
   for (size_t i = 0; i < nbyte; i++)
     BootConsPutChar(*buf++);
   return nbyte;
