+++ examples/console/tty.c
@@ -1,7 +1,7 @@
 #include "tty.h"
 #include "console.h"
 
-static int TtyWrite(File_t *f, const char *buf, size_t nbyte);
+static long TtyWrite(File_t *f, const char *buf, size_t nbyte);
 static void TtyClose(File_t *f);
 
 static FileOps_t TtyOps = {
@@ -19,7 +19,7 @@ static void TtyClose(File_t *f) {
   f->usecount--;
 }
 
-static int TtyWrite(__unused File_t *f, const char *buf, size_t nbyte) {
+static long TtyWrite(__unused File_t *f, const char *buf, size_t nbyte) {
   for (size_t i = 0; i < nbyte; i++)
     ConsolePutChar(*buf++);
   return nbyte;
